1.增加mp3 track；2.rtmp支持mp3

+++ Src/GB28181/GB28181Connection.h
@@ -37,7 +37,7 @@ public:
 
 private:
     bool _isClose = false;
-    int _ssrc = -1;
+    int64_t _ssrc = -1;
     GB28181Parser _parser;
     GB28181Context::Ptr _context;
     EventLoop::Ptr _loop;
